transformation/args/pdf_create_test.go
package args

import (
	"encoding/json"
	"testing"

	"github.com/stretchr/testify/assert"
)

func TestNewPDFCreateArgs(t *testing.T) {

	t.Run("empty map", func(t *testing.T) {

		mapExpected := map[string]interface{}{}
		expected, _ := json.Marshal(mapExpected)

		args := NewPDFCreateArgs()

		result, err := json.Marshal(args.ToMap())
		if err != nil {
			t.Error(err)
		}

		assert.Equal(t, string(expected), string(result))
	})

	t.Run("with engine", func(t *testing.T) {

		engine := "mupdf"

		mapExpected := map[string]interface{}{
			"engine": engine,
		}
		expected, _ := json.Marshal(mapExpected)

		args := NewPDFCreateArgs()
		args.SetEngine(engine)

		result, err := json.Marshal(args.ToMap())
		if err != nil {
			t.Error(err)
		}

		assert.Equal(t, string(expected), string(result))
	})

}
